+++ util/src/main/java/org/akazukin/util/time/package-info.java
@@ -4,4 +4,4 @@
  * @author currypan1229
  * @since 1.0.0
  */
-package org.akazukin.util.utils.time;
+package org.akazukin.util.time;
