+++ util/src/main/java/org/akazukin/util/time/scheduler/AScheduler.java
@@ -1,11 +1,11 @@
-package org.akazukin.util.utils.time.scheduler;
+package org.akazukin.util.time.scheduler;
 
 import lombok.AccessLevel;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 import org.akazukin.annotation.marker.ThreadSafe;
 import org.akazukin.util.object.Pair;
-import org.akazukin.util.object.TimeHolder;
+import org.akazukin.util.time.TimeHolder;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
