+++ util/src/main/java/org/akazukin/util/time/scheduler/package-info.java
@@ -4,7 +4,7 @@
  * <p>
  * Tasks are uniquely identified by numeric IDs and
  * can be scheduled with configurable delays and intervals
- * using {@link org.akazukin.util.object.TimeHolder} for time specification.
+ * using {@link org.akazukin.util.time.TimeHolder} for time specification.
  * All scheduler implementations provide exception handling
  * mechanisms through configurable {@link java.util.function.Consumer} instances
  * and support graceful shutdown for lifecycle management.
@@ -12,4 +12,4 @@
  * The scheduling infrastructure automatically handles task
  * lifecycle events, including cancellation, completion tracking, and resource cleanup.
  */
-package org.akazukin.util.utils.time.scheduler;
+package org.akazukin.util.time.scheduler;
